+++ src/components/componentInjector/componentInjector.directive.ts
@@ -1,0 +1,9 @@
+import {Directive, ViewContainerRef} from '@angular/core'
+
+@Directive({
+	selector: '[component-injector-area]'
+})
+export class ComponentInjectorAreaDirective {
+	constructor(public viewContainerRef: ViewContainerRef) {
+	}
+}