+++ src/components/componentInjector/index.ts
@@ -1,0 +1,2 @@
+export * from './componentInjector.component'
+export * from './componentInjector.directive'