+++ test/app/components/testComponent/testComponent.component.ts
@@ -1,0 +1,10 @@
+import {Component} from '@angular/core'
+
+@Component({
+	selector: 'test-component',
+	templateUrl: './testComponent.template.pug'
+})
+export class TestComponent {
+	constructor() {
+	}
+}