+++ test/webpack.config.js
@@ -1,5 +1,3 @@
-'use strict'
-
 const
 	AngularCompilerPlugin = require( '@ngtools/webpack' ).AngularCompilerPlugin,
 	BellOnBundlerErrorPlugin = require('bell-on-bundler-error-plugin'),
@@ -34,6 +32,11 @@ module.exports = () => {
 		},
 		module: {
 			rules: [
+				{
+					test: /\.html$/,
+					include,
+					use: ['raw-loader']
+				},
 				{
 					test: /\.pug$/,
 					include,